BUG: Fix bad integer if number of changed files is less
This change no longer relies on DD under the assumption that we count 5 to grab the value of changed files.
If the number of changed files was a 4 digit value then it would suffice, but anything less or greater would incidently take the next output from the total number of insertions
creating a multiline, inaccurate, answer.

For example, if report.txt displayed:

701 files changed, 300109 insertions(+), 56549 deletions(-)

Then the following command from:

echo egrep -o '[0-9][0-9]* files
changed' | egrep -o '[0-9][0-9]*' report.txt | dd status=none bs=1 count=5

Would return:

701
3

Instead of properly returning 701.

To resolve this bug, use the variable, result, as an array, and call on the
first output from that array while ignoring the rest and getting the exact results we need regardless of the number of changed files found and eliminating the return of a bad integer.