Update features
-- Add support for non positional commandline args such as targeting the
kernel path, tag regex, and parallelization.

-- Redo progress bar

-- Use temp files, especially if -j|--jobs is used, to handle subshell
variables insuring proper feedback since after the completion of said
subshells the stored variables would be lost in memory. The temp files
are set up to be removed after completion or if interrupted, and are
initially created at the root directory where this script is ran.

-- Added some additional condition checks for sake of sanity such as
insuring the regex used for tags is valid and actually exists, kernel
path is valid and it has Makefile for integrity.

-- Updated README